Add PRU classes as `friend`s of Socket class to restrict access
Declaring a class as `friend` will give the `friend`, access to all the
private members of the declaring class. Socket class is made private and
PRU and PRUSS are declared as friends of Socket class, so only they have
access to the class. Similarly, the constructors for PRU class are
declared private and PRUSS is declared as its `friend` so that PRU
can exist only inside the PRUSS class.